Allow custom constraints to be passed to getUserMedia()
There is currently no way to control the webcam media constraints. This PR changes that by allowing you to pass a `mediaConstraints` property in the `options` parameter of `tracking.track()`.

This would be very useful if you want, for example, to use a low-res feed for faster color tracking or if you need to pick a specific webcam out of many.

The suggested implementation is backwards-compatible.